[Slot Common] Temp cell is int32 and no more int16

+++ Slot_Common/common_mbox.c
@@ -11,4 +11,4 @@
 #endif
 
 commonAnalogdata_t gMboxADMes;     // contains analog measures
-uint16_t gMBoxTempCell;
+int32_t gMBoxTempCell;
